requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.27.1
httpx==0.27.0
tenacity==8.2.3
pydantic==1.10.14
python-dotenv==1.0.1
gradio==3.50.2
sqlalchemy==2.0.29
aiosqlite==0.19.0
pyyaml==6.0.1
pytest==8.1.1
pytest-playwright==0.5.4
playwright==1.42.0
